Provide type safe {un,}wrapping of ObjC promises
The Swift wrapper necessarily must throw away type information
internally to deal with storing values that are not class-based.
However, the existing API publicly throws away that type safety as well,
when it can be preserved.

By making the existing type-erasing initializer private and re-exposing
it in an extension that applies the appropriate type constraints, type
safety is reclaimed when going from Objective-C to Swift. Moving
`asObjCPromise` into the same constrained extension gains the same type
safety when going from Swift to Objective-C.

This likely breaks existing code that did perfectly safe translation
between bridged types (e.g., going from `Promise<String>` to
`FBLPromise<NSString>`). This functionality continues to be supported by
`init(bridging:)` and `asBridgedObjCPromise()`.

There may be some small minority of safe coercions that this breaks
which cannot be ported over to using the bridging extension. These can
still continue to be used, but will now require an additional `as?` to
get the desired type.